Switch configuration to @ConfigMapping
We plan to retire the legacy config classes approach at a future point
so let's migrate to @ConfigMapping.

I had to adjust the tests as the new config is immutable and based on
interfaces but it should be pretty straightforward.

Probably a good idea to merge it quickly as this PR is prone to
conflicts.